Merging in a million changes from Dev

+++ windows/checkUkmonLive.ps1
@@ -9,16 +9,16 @@ Set-Location $psscriptroot
 # read the inifile
 $ini=get-inicontent 'live.ini'
 
-#$keyfile=$ini['live']['keyfile'] 
-#$remotefolder=$ini['live']['remotefolder'] 
-#$localfolder=$ini['camera']['localfolder'] 
+$pylib=$ini['cleaning']['pylib']
 $confbad=$ini['cleaning']['badfolder'] 
 $rmsenv=$ini['rms']['rms_env'] 
 
-#move bad\*.xml .
-#move bad\*.jpg .
-
-Remove-Item $confbad/results.txt
+if ((test-path $confbad/results.txt) -eq 1){
+    Remove-Item $confbad/results.txt
+}
 conda activate $rmsenv
-python curateFolder.py live.ini
+$oldpypath=$env:pythonpath
+$env:pythonpath=$pylib
+python -m ufoutils.curateFolder live.ini
+$env:pythonpath=$oldpypath
 set-location $loc
